lib: introduce a `signing_factory` which hands out a `Signer` from `Settings`
This is needed to completely move `Signer` to `jj-core`. This also removes the `Signer::from_settings(...)`
method from Signer so it no longer is aware of `UserSettings` and replaces all callers of it with 
`signer_from_settings(...)`. Also move `SignBehavior` from `signing` to `settings` since that the only 
place which actually reads the value.
