+++ lib/tests/test_signing.rs
@@ -4,9 +4,9 @@ use jj_lib::backend::Timestamp;
 use jj_lib::config::ConfigLayer;
 use jj_lib::config::ConfigSource;
 use jj_lib::repo::Repo as _;
+use jj_lib::settings::SignBehavior;
 use jj_lib::settings::UserSettings;
 use jj_lib::signing::SigStatus;
-use jj_lib::signing::SignBehavior;
 use jj_lib::signing::Signer;
 use jj_lib::signing::Verification;
 use jj_lib::test_signing_backend::TestSigningBackend;
